feat(marc-record-validation) move Marc record validation to MarcRecordsService.
