+++ src/main/java/org/folio/qm/service/ValidationService.java
@@ -15,6 +15,4 @@ public interface ValidationService {
   List<ValidationIssue> validate(ValidatableRecord validatableRecord);
 
   void validateIdsMatch(QuickMarcEdit quickMarc, UUID externalId);
-
-  void validateMarcRecord(BaseMarcRecord marcRecord);
 }
